requirements.txt
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.0
python-dotenv==1.0.0
spotipy==2.23.0
ytmusicapi==1.3.0